VEMIRT: A Package for High-Dimensional IRT Models

VEMIRT is created to assist researchers to conduct exploratory and confirmatory multidimensional item response theory (MIRT) analysis and cooresponding item differential functioning (DIF) analysis. The core computation engine of VEMIRT is a family of Gaussian Variational EM algorithms that are considerably more efficient than currently available algorithms in other software packages, especially when the number of latent factors exceeds four.

Identifying the number of factors

pa_poly identifies the number of factors via parallel analysis.

Exploratory factor analysis

E2PL_gvem_rot conducts M2PL Analysis with post-hoc rotation (Promax & CF-Quartimax)

E2PL_gvem_lasso conducts M2PL Analysis with Lasso penalty

E2PL_gvem_adaptlasso conducts M2PL Analysis with adaptive Lasso penalty

E2PL_iw conducts importance sampling to correct bias for M2PL analysis

E3PL_sgvem_rot conducts stochastic GVEM to further improve the computational effficiency for exploratory M3PL analysis

E3PL_sgvem_lasso conducts M3PL Analysis with Lasso penalty

E3PL_sgvem_adaptlasso conducts M3PL Analysis with adaptive Lasso penalty

MGRM_gvem conducts GVEM for the multidimensional graded response model with post-hoc rotation

MGPCM_gvem conducts GVEM for the multidimensional partial credit model with post-hoc rotation

Confirmatory factor analysis

C2PL_gvem conducts GVEM for confirmatory M2PL analysis

C2PL_bs conducts bootstrap sampling to correct bias and produce standard errors for confirmatory M2PL analysis

C2PL_iw conducts importance sampling to correct bias for M2PL analysis

C2PL_iw2 conducts IW-GVEM for confirmatory M2PL analysis (alternative implementation to C2PL_iw)

C3PL_sgvem conducts stochastic GVEM for confirmatory M3PL analysis

MGRM_gvem conducts GVEM for the multidimensional graded response model

MGPCM_gvem conducts GVEM for the multidimensional partial credit model

Differential item functioning analysis

D1PL_em conducts DIF analysis for M1PL models using EM algorithms

D1PL_gvem conducts DIF analysis for M1PL models using GVEM algorithms

D2PL_em conducts DIF analysis for M2PL models using EM algorithms

D2PL_pair_em conducts DIF analysis for 2PL models using EM algorithms with group pairwise truncated \(L_1\) penalty

D2PL_gvem conducts DIF analysis for M2PL models using GVEM algorithms

D2PL_lrt conducts DIF analysis for M2PL models using the likelihood ratio test

Shiny apps for VEMIRT

shinyVEMIRT Run the shiny app for VEMIRT

DIFdashboard Run the shiny app for DIF Dashboard

See also

Author

Maintainer: Weicong Lyu weiconglyu@um.edu.mo (ORCID)

Authors:

Yijun Cheng chengxb@uw.edu (ORCID)

Jiaying Xiao jxiao6@uw.edu (ORCID)

He Ren heren@uw.edu (ORCID)

Ruoyi Zhu zhux0445@uw.edu (ORCID)

Gongjun Xu gongjun@umich.edu (ORCID)

Chun Wang wang4066@uw.edu (ORCID)