VEMIRT: A package for high-dimensional IRT models

VEMIRT is created to assist researchers to conduct exploratory and confirmatory multidimensional item response theory (MIRT) analysis and cooresponding item differential functioning (DIF) analysis. The core computation engine of VEMIRT is a family of Gaussian Variational EM algorithms that are considerably more efficient than currently available algorithms in other software packages, especially when the number of latent factors exceeds four.

Identifying the number of factors

pa_poly identifies the number of factors via parallel analysis.

Exploratory factor analysis

E2PL_gvem_rot conducts M2PL Analysis with post-hoc rotation (Promax & CF-Quartimax)

E2PL_gvem_lasso conducts M2PL Analysis with Lasso penalty

E2PL_gvem_adaptlasso conducts M2PL Analysis with adaptive Lasso penalty

E2PL_iw conducts importance sampling to correct bias for M2PL analysis

E3PL_sgvem_rot conducts stochastic GVEM to futher imporve the computational effficiency for exploratory M3PL analysis

E3PL_sgvem_lasso conducts M3PL Analysis with Lasso penalty

E3PL_sgvem_adaptlasso conducts M3PL Analysis with adaptive Lasso penalty

Confirmatory factor analysis

C2PL_gvem conducts GVEM for confirmatory M2PL analysis

C2PL_bs conducts bootstrap sampling to correct bias and produce standard errors for confirmatory M2PL analysis

C2PL_iw conducts importance sampling to correct bias for M2PL analysis

C3PL_sgvem conducts stochastic GVEM for confirmatory M3PL analysis

Differential item functioning analysis

D2PL_em conducts DIF analysis for M2PL models using EM algorithms

D2PL_gvem conducts DIF analysis for M2PL models using GVEM algorithms

D2PL_lrt conducts DIF analysis for M2PL models using the likelihood ratio test

See also

Author

Maintainer: Weicong Lyu wlyu4@uw.edu (ORCID)

Authors:

Jiaying Xiao jxiao6@uw.edu (ORCID)

Ruoyi Zhu zhux0445@uw.edu (ORCID)

Gongjun Xu gongjun@umich.edu (ORCID)

Chun Wang wang4066@uw.edu (ORCID)